test: flatten standard-schema spec into a single file
