+++ tests/index.ts
@@ -4,5 +4,5 @@ describe('type-flag', () => {
 	import('./specs/type-flag/index.ts');
 	import('./specs/get-flag/index.ts');
 	import('./specs/flag-name-to-kebab/index.ts');
-	import('./specs/standard-schema/index.ts');
+	import('./specs/standard-schema.ts');
 });
