--- tests/specs/standard-schema/index.ts
@@ -1,5 +1,0 @@
-import { describe } from 'manten';
-
-describe('standard-schema', () => {
-	import('./parsing.ts');
-});
